requirements.txt
connexion >= 2.6.0
python_dateutil >= 2.6.0
setuptools >= 21.0.0

flask-cors == 3.0.9

beautifulsoup4>=4.9.3
lxml>=4.6.2
mecab-python3
pandas == 1.1.5
gensim == 3.6.0
scikit-learn == 0.22.2.post1

lightgbm == 2.2.3
